+++ packages/amplify-category-api/src/index.ts
@@ -3,6 +3,7 @@ import {
   $TSContext,
   $TSObject,
   AmplifyCategories,
+  AmplifyError,
   AmplifySupportedService,
   buildOverrideDir,
   pathManager,
@@ -27,6 +28,7 @@ import { getAPIResourceDir } from './provider-utils/awscloudformation/utils/ampl
 import { configureMultiEnvDBSecrets } from './provider-utils/awscloudformation/utils/rds-secrets/multi-env-database-secrets';
 import { deleteConnectionSecrets, readDatabaseNameFromMeta } from './provider-utils/awscloudformation/utils/rds-secrets/database-secrets';
 import _ from 'lodash';
+import { AmplifyGraphQLTransformerErrorConverter } from './errors/amplify-error-converter';
 
 export { NETWORK_STACK_LOGICAL_ID } from './category-constants';
 export { addAdminQueriesApi, updateAdminQueriesApi } from './provider-utils/awscloudformation';
@@ -349,18 +351,27 @@ export const transformCategoryStack = async (context: $TSContext, resource: $TSO
     if (canResourceBeTransformed(resource.resourceName)) {
       const backendDir = pathManager.getBackendDirPath();
       const overrideDir = path.join(backendDir, resource.category, resource.resourceName);
-      const isBuild = await buildOverrideDir(backendDir, overrideDir);
-      await context.amplify.invokePluginMethod(context, 'awscloudformation', undefined, 'compileSchema', [
-        context,
-        {
-          forceCompile: true,
-          overrideConfig: {
-            overrideFlag: isBuild,
-            overrideDir,
-            resourceName: resource.resourceName,
+      const isBuild = await buildOverrideDir(backendDir, overrideDir).catch(error => {
+        throw new AmplifyError('InvalidOverrideError', {
+          message: error.message,
+          link: 'https://docs.amplify.aws/cli/graphql/override/',
+        });
+      });
+      try {
+        await context.amplify.invokePluginMethod(context, 'awscloudformation', undefined, 'compileSchema', [
+          context,
+          {
+            forceCompile: true,
+            overrideConfig: {
+              overrideFlag: isBuild,
+              overrideDir,
+              resourceName: resource.resourceName,
+            },
           },
-        },
-      ]);
+        ]);
+      } catch (error) {
+        throw AmplifyGraphQLTransformerErrorConverter.convert(error);
+      }
     }
   } else if (resource.service === AmplifySupportedService.APIGW) {
     if (canResourceBeTransformed(resource.resourceName)) {