+++ packages/amplify-graphql-transformer-interfaces/src/transform-host-provider.ts
@@ -6,7 +6,7 @@ import {
   HttpDataSource,
   LambdaDataSource,
   NoneDataSource,
-} from '@aws-cdk/aws-appsync-alpha';
+} from 'aws-cdk-lib/aws-appsync';
 import { CfnResolver } from 'aws-cdk-lib/aws-appsync';
 import { ITable } from 'aws-cdk-lib/aws-dynamodb';
 import { IFunction, ILayerVersion, Runtime } from 'aws-cdk-lib/aws-lambda';