+++ packages/amplify-graphql-http-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-http-transformer@1.1.0...@aws-amplify/graphql-http-transformer@1.1.1) (2023-03-15)
+
+**Note:** Version bump only for package @aws-amplify/graphql-http-transformer
+
 # [1.1.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-http-transformer@0.8.37...@aws-amplify/graphql-http-transformer@1.1.0) (2023-03-01)
 
 ### Bug Fixes