+++ packages/amplify-migration-tests/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.19](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-migration-tests@5.0.18...amplify-category-api-migration-tests@5.0.19) (2023-03-15)
+
+**Note:** Version bump only for package amplify-category-api-migration-tests
+
 ## [5.0.18](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-migration-tests@5.0.17...amplify-category-api-migration-tests@5.0.18) (2023-03-01)
 
 **Note:** Version bump only for package amplify-category-api-migration-tests